[Front] migration of Reports.js to tsx pure function (#2588)
